benchmark/requirements.txt
torch>=2.1
pyyaml>=6.0
tensorboard>=2.14
h5py>=3.9
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
# Optional: for MeshGraphNet
# torch-geometric>=2.4
